Version 1.0
First release of WikiWallpaper. Added install instructions, and .exe for use on any Windows 10 Computer. Made a number of tweaks to make the program run the same in an IDE as it runs as a .jar and .exe file.  The java program will exit when the window closes now to stop run on tasks. Icons and Logo are not stored in the cloud (imgur) and download when the .jar is first run.